Starting Qt version
I don't know yet where this is going to lead. For now, porting Free42
is just an obvious way to learn about Qt, but it is conceivable that
this may end up replacing the GTK version, and who knows, maybe some or
all of the other versions as well. Having a cross-platform UI would
certainly be nice in that doing UI enhancements once would beat doing
them five times.